Fix photo upload (#13)

+++ routes/admin/articles.js
@@ -33,7 +33,7 @@ function handlePicture(req, article, callback) {
         secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
         region: process.env.AWS_S3_BUCKET_REGION
       });
-      if (article) {
+      if (article && article.pictureUrl && article.pictureUrl != '') {
         //// delete existing picture, if any
         s3.deleteObject({
           Bucket: process.env.AWS_S3_BUCKET,
